pyproject.toml
[project]
name = "optic-mcp"
version = "0.3.0"
description = "MCP server for camera/vision tools - connect any camera to AI assistants"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Timor Leiderman", email = "timorleiderman@gmail.com" }
]
keywords = ["mcp", "camera", "opencv", "ai", "llm", "vision", "usb-camera", "model-context-protocol", "ip-camera", "rtsp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video :: Capture",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "opencv-python>=4.8.0",
    "requests>=2.28.0",
    "mss>=9.0.0",
    "Pillow>=10.0.0",
    "pyzbar>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "numpy>=1.24.0",
    "ruff>=0.8.0",
    "twine>=5.0.0",
    "bump2version>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Timorleiderman/OpticMCP"
Repository = "https://github.com/Timorleiderman/OpticMCP"
Issues = "https://github.com/Timorleiderman/OpticMCP/issues"

[project.scripts]
optic-mcp = "optic_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/optic_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
line-length = 100
